refactor: favorites.controller

+++ src/favorites/favorites.controller.ts
@@ -1,3 +1,4 @@
+import { NotFoundException } from '@nestjs/common'
 import {
   ClassSerializerInterceptor,
   Controller,
@@ -11,13 +12,14 @@ import {
   Req,
   UseInterceptors,
 } from '@nestjs/common'
+import { Request } from 'express'
 import { FavoritesService } from './favorites.service'
 
+type Path = 'artist' | 'album' | 'track'
 
 @Controller('favs')
 export class FavoritesController {
   constructor(private readonly favoriteService: FavoritesService) {}
-
 
   @UseInterceptors(ClassSerializerInterceptor)
   @Get()
@@ -27,14 +29,37 @@ export class FavoritesController {
 
   @Post(':path/:id')
   @HttpCode(HttpStatus.CREATED)
-  addArtist(@Param('path') path:  'artist' | 'album' | 'track', @Param('id', ParseUUIDPipe) id: string) {
-    return this.favoriteService.add(id, path)
+  addArtist(
+    @Req() req: Request, 
+    @Param('path') path: Path, 
+    @Param('id', ParseUUIDPipe) id: string,
+  ) {
+    if (path === 'artist' || path === 'album' || path === 'track') {
+      return this.favoriteService.add(id, path)
+    } else {
+      throw new NotFoundException({ 
+        statusCode: 404, 
+        message: `Cannot POST ${req.originalUrl}`, 
+        error: 'Not Found'
+      })
+    }
   }
 
   @Delete(':path/:id')
   @HttpCode(HttpStatus.NO_CONTENT)
-  deleteArtist(@Param('path') path: 'artist' | 'album' | 'track', @Param('id', ParseUUIDPipe) id: string) {
-    return this.favoriteService.delete(id, path)
+  deleteArtist(
+    @Req() req: Request,
+    @Param('path') path: Path,
+    @Param('id', ParseUUIDPipe) id: string,
+  ) {
+    if (path === 'artist' || path === 'album' || path === 'track') {
+      return this.favoriteService.delete(id, path)
+    } else {
+      throw new NotFoundException({ 
+        statusCode: 404, 
+        message: `Cannot DELETE ${req.originalUrl}`, 
+        error: 'Not Found' 
+      })
+    }
   }
-
 }
